refactor: cartList, PaymentsAmount Props 수정
